+++ packages/app-shell/src/Confirmation/getCopy.test.jsx
@@ -2,7 +2,7 @@ import getCopy, { SUCCESS_CTA } from './getCopy';
 import { renderHook } from '@testing-library/react-hooks';
 
 describe('getCopy', () => {
-  it('should return label, description and buttonCopy for a team/individual plan change', () => {
+  it('should return label, description and buttonCopy for a team/essentials plan change', () => {
     const planName = 'Team';
 
     const { result } = renderHook(() => getCopy({ planName }));
@@ -44,12 +44,10 @@ describe('getCopy', () => {
 
     const { result } = renderHook(() => getCopy({ planName, startedTrial }));
 
-    expect(result.current.label).toBe(
-      'Congrats! You are now starting your trial'
-    );
+    expect(result.current.label).toBe('Trial activated! Time to explore.');
     expect(result.current.description).toBe(
-      'For the next 14 days you get to experience Buffer to it’s full. Have fun!'
+      'Let’s make the most of your 14-day trial. Jump in and start exploring your advanced publishing, analytics, and engagement features. '
     );
-    expect(result.current.buttonCopy).toBe(SUCCESS_CTA);
+    expect(result.current.buttonCopy).toBe("Let's Go!");
   });
 });