+++ packages/app-shell/src/Modal/index.jsx
@@ -9,32 +9,47 @@ import Confirmation from '../Confirmation';
 const ModalContent = ({ modal }) => {
   switch (modal) {
     case MODALS.paymentMethod:
-      return (<PaymentMethod />);
+      return <PaymentMethod />;
     case MODALS.planSelector:
-      return (<PlanSelector />);
+      return <PlanSelector />;
     case MODALS.success:
-      return (<Confirmation />);
+      return <Confirmation />;
     case MODALS.startTrial:
-      return (<StartTrial />);
+      return <StartTrial />;
     default:
       return null;
   }
-}
+};
 
-const Modal = ({ modal, openModal }) => {
+const Modal = ({ modal, openModal, isAwaitingUserAction }) => {
   const [hasModal, setHasModal] = useState(!!modal);
+
+  useEffect(() => {
+    if (isAwaitingUserAction) {
+      openModal(MODALS.planSelector, {
+        cta: 'awaitingUserAction',
+        isUpgradeIntent: false,
+      });
+    }
+  }, [isAwaitingUserAction]);
 
   useEffect(() => {
     setHasModal(!!modal);
   }, [modal]);
 
-  return (<>
-    {hasModal && <SimpleModal closeAction={() => {
-      openModal(null);
-    }}>
-      <ModalContent modal={modal} />
-    </SimpleModal>}
-  </>)
-}
+  return (
+    <>
+      {hasModal && (
+        <SimpleModal
+          closeAction={() => {
+            openModal(null);
+          }}
+        >
+          <ModalContent modal={modal} />
+        </SimpleModal>
+      )}
+    </>
+  );
+};
 
 export default Modal;