+++ packages/app-shell/src/PaymentMethod/hooks/useSetupIntent.test.jsx
@@ -5,12 +5,17 @@ import { renderHook } from '@testing-library/react-hooks'
 import { CREATE_SETUP_INTENT } from '../../graphql/billing'
 import useSetupIntent from './useSetupIntent'
 
+
+
 describe('useSetupIntent', () => {
   const mockMutation = jest.fn(() => ({
     data: {
-      billingCreateSetupIntent: 'fooSetupIntent',
-    }
-  }))
+      billingCreateSetupIntent: {
+        success: true,
+        clientSecret: 'fooSetupIntent',
+      },
+    },
+  }));
   const user = {
     currentOrganization: {
       id: '123FooOrganization',