+++ packages/app-shell/src/PlanSelector/hooks/useSelectedPlan.test.jsx
@@ -2,7 +2,7 @@ import useSelectedPlan from './useSelectedPlan';
 import { renderHook, act } from '@testing-library/react-hooks';
 
 const planOptions = [
-  { planId: 'individual', isCurrentPlan: false, planInterval: 'year' },
+  { planId: 'essentials', isCurrentPlan: false, planInterval: 'year' },
   { planId: 'team', isCurrentPlan: true, planInterval: 'year' },
   { planId: 'free', planInterval: 'month' },
 ];
@@ -12,7 +12,7 @@ describe('useSelectedPlan', () => {
     const { result } = renderHook(() => useSelectedPlan(planOptions));
     expect(result.current.selectedPlan.planId).toBe('team');
   });
-  it('should set the default as the individual plan if the intent is to upgrade', () => {
+  it('should set the default as the essentials plan if the intent is to upgrade', () => {
     const isUpgradeIntent = true;
 
     const { result } = renderHook(() =>
@@ -21,13 +21,13 @@ describe('useSelectedPlan', () => {
     expect(result.current.selectedPlan.planId).toBe('team');
   });
   it('should update the selected plan', () => {
-    const planString = 'individual_year';
+    const planString = 'essentials_year';
 
     const { result } = renderHook(() => useSelectedPlan(planOptions));
     act(() => {
       result.current.updateSelectedPlan(planString);
     });
-    expect(result.current.selectedPlan.planId).toBe('individual');
+    expect(result.current.selectedPlan.planId).toBe('essentials');
   });
   it('should update the selected plan with a new interval', () => {
     const planString = 'free_month';