+++ packages/app-shell/src/StartTrial/index.jsx
@@ -1,98 +1,121 @@
 import React, { useEffect, useState } from 'react';
-import { useMutation } from '@apollo/client';
 
 import CheckmarkIcon from '@bufferapp/ui/Icon/Icons/Checkmark';
 import Text from '@bufferapp/ui/Text';
 import Button from '@bufferapp/ui/Button';
 
-import { Error } from '../PaymentMethod/style'
+import { Error } from '../PaymentMethod/style';
 import { MODALS } from '../hooks/useModal';
 import { UserContext } from '../context/User';
 import { ModalContext } from '../context/Modal';
-import { START_TRIAL } from '../graphql/billing';
-import { QUERY_ACCOUNT } from '../graphql/account'
+import useStartTrial from '../hooks/useStartTrial';
 
-import {
-  Holder,
-  Content,
-  Ctas,
-} from './style'
+import { Holder, Content, Ctas } from './style';
 
-const StartTrial = ({ user, openModal}) => {
-  const [suggestedPlan, setSuggestedPlan] = useState(null)
-  const [processing, setProcessing] = useState(false)
+const StartTrial = ({ user, openModal }) => {
+  const [suggestedPlan, setSuggestedPlan] = useState(null);
   useEffect(() => {
     if (user) {
-      let plan = user.currentOrganization.billing.changePlanOptions.find(p => p.isRecommended)
+      let plan = user.currentOrganization?.billing?.changePlanOptions.find(
+        (p) => p.isRecommended
+      );
       if (!plan) {
         plan = {
           planId: 'team',
-          planInterval: "month",
-        }
+          planInterval: 'month',
+        };
       }
-      setSuggestedPlan(plan)
+      setSuggestedPlan(plan);
     }
-  }, [user])
+  }, [user]);
 
-  const [startTrial, { data:trial, error }] = useMutation(
-    START_TRIAL,
-    {
-      refetchQueries: [{ query: QUERY_ACCOUNT }],
-    }
-  );
+  const { startTrial, trial, error, processing } = useStartTrial({
+    user,
+    plan: suggestedPlan,
+  });
 
   useEffect(() => {
-    if (trial) {
-      openModal(MODALS.success, { startedTrial: true })
+    if (trial && trial.billingStartTrial.success) {
+      openModal(MODALS.success, { startedTrial: true });
     }
-  }, [trial])
+  }, [trial]);
 
-  return(<Holder>
-    <Content>
-      <Text type="h1">Want to try our trial?</Text>
-      <Text type="p">Get the best we have to offer to see how it fits for you and your business.</Text>
-      <ol>
-        <li> <CheckmarkIcon size="medium" /><Text>Unlimited users</Text></li>
-        <li> <CheckmarkIcon size="medium" /><Text>Unlimited channels</Text></li>
-        <li> <CheckmarkIcon size="medium" /><Text>No credit card required</Text></li>
-      </ol>
-      <Ctas>
-        <Button
-          type="primary"
-          disabled={!suggestedPlan || processing}
-          onClick={() => {
-            setProcessing(true)
-            startTrial({
-              variables: {
-                organizationId: user.currentOrganization.id,
-                plan: suggestedPlan.planId,
-                interval: suggestedPlan.planInterval,
-              },
-            }).catch((e) => {
-              setProcessing(false)
-              console.error(e)
-            });
-          }}
-          label={processing ? "Processing ..." : "Start Free 14-day Trial"}
+  return (
+    <Holder>
+      <Content>
+        <Text type="h1">Test run our paid features</Text>
+        <Text type="p">
+          Get faster results from your creative with our best publishing,
+          analytics, and engagement tools built for growing businesses.
+        </Text>
+        <ol>
+          <li>
+            {' '}
+            <CheckmarkIcon size="medium" />
+            <Text>Advanced Instagram features</Text>
+          </li>
+          <li>
+            {' '}
+            <CheckmarkIcon size="medium" />
+            <Text>Machine-learning driven insights</Text>
+          </li>
+          <li>
+            {' '}
+            <CheckmarkIcon size="medium" />
+            <Text>
+              Analytics for the best time to post and your audience demographics
+            </Text>
+          </li>
+          <li>
+            {' '}
+            <CheckmarkIcon size="medium" />
+            <Text>Easy, automatic reporting on growth and engagement</Text>
+          </li>
+        </ol>
+        <Ctas>
+          <Button
+            type="secondary"
+            onClick={() => {
+              openModal(MODALS.planSelector, {
+                cta: 'planSelection',
+                ctaButton: 'checkOutPaidPlans',
+              });
+            }}
+            label="I'm ready to upgrade"
+          />
+          <Button
+            type="primary"
+            disabled={!suggestedPlan || processing}
+            onClick={() => {
+              startTrial();
+            }}
+            label={processing ? 'Processing ...' : 'Start my 14-day Free Trial'}
+          />
+        </Ctas>
+        <Error
+          error={
+            error
+              ? {
+                  message: error.message,
+                }
+              : null
+          }
         />
-        <Button
-          type="secondary"
-          onClick={() => {openModal(MODALS.planSelector)}}
-          label="Check Out Paid Plans"
-        />
-      </Ctas>
-      <Error error={error ? { message: "We can't start your trial, please contact support." } : null} />
-    </Content>
-  </Holder>)
-}
+      </Content>
+    </Holder>
+  );
+};
 
 const StartTrialProvider = () => {
-  return (<UserContext.Consumer>
-    {user => (<ModalContext.Consumer>{
-      ({ openModal }) => (
-        <StartTrial user={user} openModal={openModal} />
-      )}</ModalContext.Consumer>)}
-  </UserContext.Consumer>)
-}
+  return (
+    <UserContext.Consumer>
+      {(user) => (
+        <ModalContext.Consumer>
+          {({ openModal }) => <StartTrial user={user} openModal={openModal} />}
+        </ModalContext.Consumer>
+      )}
+    </UserContext.Consumer>
+  );
+};
 
-export default StartTrialProvider
+export default StartTrialProvider;