+++ packages/app-shell/src/Summary/index.jsx
@@ -17,7 +17,6 @@ import {
   SummaryNote,
 } from './style';
 import { UserContext } from '../context/User';
-import { freePlan } from '../mocks/freePlan';
 
 const Summary = ({
   planOptions,
@@ -34,7 +33,7 @@ const Summary = ({
     : '';
 
   const isDowngrading = (currentPlanId, selectedPlanId) => {
-    if (currentPlanId === 'individual') {
+    if (currentPlanId === 'essentials') {
       return selectedPlanId === 'free' ? true : false;
     }
     if (currentPlanId === 'team') {
@@ -96,7 +95,7 @@ const Summary = ({
         return null;
       }
       return (
-        <Text type="label" color="grayDark">
+        <Text type="label" color="grayDarker">
           {/* this ends up reading: # social channels x base price */}
           {`${selectedPlan.channelsQuantity} social channel${
             selectedPlan.channelsQuantity > 1 ? 's' : ''
@@ -111,7 +110,7 @@ const Summary = ({
       );
     } else {
       return (
-        <Text type="label" color="grayDark">
+        <Text type="label" color="grayDarker">
           Includes tax
         </Text>
       );
@@ -135,6 +134,26 @@ const Summary = ({
   const intervalInWords =
     selectedPlan.planInterval === 'month' ? '30 days' : 'year';
 
+  const getSummaryNote = () => {
+    if (trialInfo?.isActive) {
+      return (
+        <Text type="p">
+          You won't be charged until the end of your trial on{' '}
+          <span>{formattedTrialEndDate}</span>
+        </Text>
+      );
+    } else if (selectedPlan.planId === 'free' && subscriptionEndDate) {
+      return (
+        <Text type="p">
+          Changing to Free will occur at the end of your next billing cycle on{' '}
+          <span>{formattedSubscriptionEndDate}</span>
+        </Text>
+      );
+    } else if (selectedPlan.planId === 'free' && !subscriptionEndDate) {
+      return <Text type="p">Upgrade your plan at anytime</Text>;
+    } else return <Text type="p">Cancel your plan at anytime</Text>;
+  };
+
   return (
     <SummaryContainer>
       <Body>
@@ -150,16 +169,7 @@ const Summary = ({
               ))}
             </DetailList>
             <Separator />
-            <SummaryNote>
-              {selectedPlan.planId === 'free' ? (
-                <Text type="p">
-                  Changing to Free will occur at the end of your next billing
-                  cycle on <span>{formattedSubscriptionEndDate}</span>
-                </Text>
-              ) : (
-                <Text type="p">Cancel your plan at anytime</Text>
-              )}
-            </SummaryNote>
+            <SummaryNote>{getSummaryNote()}</SummaryNote>
           </>
         ) : (
           <>
@@ -174,7 +184,7 @@ const Summary = ({
               {trialInfo?.isActive ? (
                 <Text type="p">
                   You won't be charged until the end of your trial on{' '}
-                  <span>{formattedTrialEndDate}</span>
+                  <b>{formattedTrialEndDate}</b>
                 </Text>
               ) : (
                 <Text type="p">