+++ packages/app-shell/src/graphql/account.js
@@ -27,6 +27,7 @@ export const QUERY_ACCOUNT = gql`
         role
         createdAt
         isOneBufferOrganization
+        featureFlips
         billing {
           canAccessAnalytics
           canAccessEngagement
@@ -36,6 +37,10 @@ export const QUERY_ACCOUNT = gql`
           }
           ... on MPBilling {
             billingRedirectUrl
+            subscriptions {
+              plan
+              product
+            }
           }
           ... on OBBilling {
             canStartTrial
@@ -50,6 +55,7 @@ export const QUERY_ACCOUNT = gql`
               }
               plan {
                 id
+                name
               }
             }
             changePlanOptions {
@@ -85,6 +91,7 @@ export const QUERY_ACCOUNT = gql`
         role
         createdAt
         isOneBufferOrganization
+        featureFlips
         billing {
           canAccessAnalytics
           canAccessEngagement
@@ -94,6 +101,10 @@ export const QUERY_ACCOUNT = gql`
           }
           ... on MPBilling {
             billingRedirectUrl
+            subscriptions {
+              plan
+              product
+            }
           }
           ... on OBBilling {
             canStartTrial
@@ -108,6 +119,7 @@ export const QUERY_ACCOUNT = gql`
               }
               plan {
                 id
+                name
               }
             }
             changePlanOptions {